refactor: complete codebase cleanup
