+++ app/src/main/java/dev/hazydreams/hermesceleste/ui/sessions/SessionListScreen.kt
@@ -1,0 +1,166 @@
+package dev.hazydreams.hermesceleste.ui.sessions
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.DropdownMenu
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import dev.hazydreams.hermesceleste.network.DashboardProfile
+import dev.hazydreams.hermesceleste.network.StoredSession
+import dev.hazydreams.hermesceleste.ui.CelesteBlue
+import dev.hazydreams.hermesceleste.ui.CelesteError
+import dev.hazydreams.hermesceleste.ui.CelesteMuted
+import dev.hazydreams.hermesceleste.ui.CelestePanel
+import dev.hazydreams.hermesceleste.ui.CelesteBackdrop
+import dev.hazydreams.hermesceleste.ui.StatusMessage
+
+@Composable
+internal fun SessionListScreen(
+    sessions: List<StoredSession>,
+    profiles: List<DashboardProfile>,
+    selectedProfile: String,
+    loadingMessage: String?,
+    errorMessage: String?,
+    onProfileSelected: (String) -> Unit,
+    onNewConversation: () -> Unit,
+    onSessionSelected: (StoredSession) -> Unit,
+    onSettings: () -> Unit,
+) {
+    var profileMenuExpanded by remember { mutableStateOf(false) }
+    CelesteBackdrop {
+        Column(Modifier.fillMaxSize().padding(top = 46.dp)) {
+            Row(
+                modifier = Modifier.fillMaxWidth().padding(horizontal = 20.dp),
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.spacedBy(14.dp),
+            ) {
+                Column(Modifier.weight(1f)) {
+                    Text("Conversations", fontSize = 26.sp, fontWeight = FontWeight.SemiBold)
+                    Text("${sessions.size} from this Hermes", color = CelesteMuted, fontSize = 13.sp)
+                }
+                TextButton(
+                    onClick = onSettings,
+                    enabled = loadingMessage == null,
+                ) {
+                    Text("Settings")
+                }
+            }
+            Row(
+                modifier = Modifier.fillMaxWidth().padding(horizontal = 20.dp, vertical = 12.dp),
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.spacedBy(10.dp),
+            ) {
+                Box(Modifier.weight(1f)) {
+                    OutlinedButton(
+                        onClick = { profileMenuExpanded = true },
+                        enabled = loadingMessage == null,
+                        modifier = Modifier.fillMaxWidth(),
+                        shape = RoundedCornerShape(14.dp),
+                    ) {
+                        Text("Profile: ${selectedProfile.replaceFirstChar(Char::uppercase)}")
+                    }
+                    DropdownMenu(
+                        expanded = profileMenuExpanded,
+                        onDismissRequest = { profileMenuExpanded = false },
+                    ) {
+                        profiles.forEach { profile ->
+                            DropdownMenuItem(
+                                text = {
+                                    Text(
+                                        if (profile.isDefault) "${profile.name} · default" else profile.name,
+                                    )
+                                },
+                                onClick = {
+                                    onProfileSelected(profile.name)
+                                    profileMenuExpanded = false
+                                },
+                            )
+                        }
+                    }
+                }
+                Button(
+                    onClick = onNewConversation,
+                    enabled = loadingMessage == null,
+                    shape = RoundedCornerShape(14.dp),
+                ) {
+                    Text("New chat")
+                }
+            }
+            if (loadingMessage != null || errorMessage != null) {
+                Column(Modifier.padding(horizontal = 20.dp, vertical = 12.dp)) {
+                    loadingMessage?.let { StatusMessage(it, CelesteBlue, showSpinner = true) }
+                    errorMessage?.let { StatusMessage(it, CelesteError) }
+                }
+            }
+            LazyColumn(
+                contentPadding = PaddingValues(horizontal = 20.dp, vertical = 18.dp),
+                verticalArrangement = Arrangement.spacedBy(10.dp),
+            ) {
+                items(sessions, key = { it.id }) { session ->
+                    Column(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .background(CelestePanel.copy(alpha = 0.94f), RoundedCornerShape(20.dp))
+                            .border(1.dp, Color.White.copy(alpha = 0.07f), RoundedCornerShape(20.dp))
+                            .clickable(enabled = loadingMessage == null) { onSessionSelected(session) }
+                            .padding(18.dp),
+                    ) {
+                        Text(
+                            session.title.ifBlank { "Untitled conversation" },
+                            fontSize = 16.sp,
+                            fontWeight = FontWeight.SemiBold,
+                            maxLines = 1,
+                            overflow = TextOverflow.Ellipsis,
+                        )
+                        if (session.preview.isNotBlank()) {
+                            Spacer(Modifier.height(7.dp))
+                            Text(
+                                session.preview,
+                                color = CelesteMuted,
+                                fontSize = 13.sp,
+                                lineHeight = 19.sp,
+                                maxLines = 2,
+                                overflow = TextOverflow.Ellipsis,
+                            )
+                        }
+                        Spacer(Modifier.height(10.dp))
+                        Text(
+                            "${session.messageCount} messages  •  ${session.profile}  •  ${session.source.ifBlank { "Hermes" }}",
+                            color = CelesteBlue.copy(alpha = 0.86f),
+                            fontSize = 11.sp,
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
